+++ core/src/main/java/com/sopt/core/designsystem/component/calendar/CalendarWeek.kt
@@ -13,6 +13,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.sopt.core.designsystem.component.calendar.CalendarDay
+import com.sopt.core.designsystem.component.calendar.CalendarScheduleGroup
 import com.sopt.core.designsystem.theme.NoostakAndroidTheme
 import com.sopt.core.designsystem.theme.NoostakTheme
 import com.sopt.core.extension.isToday
